requirements.txt
python-dateutil==2.8.2
streamlit==1.21.0
pandas==1.5.3
streamlit-folium==0.11.1
Pillow==9.5.0
openpyxl==3.1.2
